requirements.txt
streamlit==1.23.1
st-audiorec==0.0.2
openai==0.27.7
python-dotenv==1.0.0
